requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
python-dotenv
google-generativeai